# api-documentation-old.yaml
openapi: 3.0.0
info:
  title: InCrowd API
  version: 1.0.0
  description: |
    # InCrowd API - Documentazione Completa
    
    Benvenuto nella documentazione dell'API di InCrowd, una piattaforma innovativa di crowdsourcing per proposte e iniziative cittadine.
    
    ## Panoramica
    
    L'API InCrowd fornisce accesso completo alla piattaforma per:
    - Gestione utenti (cittadini privati ed enti)
    - Creazione e gestione proposte
    - Sistema di follow/followers
    - Amministrazione e moderazione
    
    ## Autenticazione e Sicurezza
    
    L'API utilizza autenticazione JWT (JSON Web Token):
    
    1. **Login**: Ottieni il token tramite `POST /api/auth/login`
    2. **Utilizzo**: Includi il token nell'header di ogni richiesta:
       ```
       Authorization: Bearer <your-jwt-token>
       ```
    3. **Scadenza**: I token hanno durata limitata, effettua refresh quando necessario
    
    ## Struttura delle Response
    
    Tutte le risposte seguono il formato standardizzato:
    
    ```json
    {
      "data": { /* Dati della risposta */ },
      "message": "Messaggio descrittivo",
      "error": null // Solo presente in caso di errori
    }
    ```
    
    ## Gestione Errori
    
    - **400 Bad Request**: Dati non validi o parametri mancanti
    - **401 Unauthorized**: Token mancante, scaduto o non valido
    - **403 Forbidden**: Permessi insufficienti per l'operazione
    - **404 Not Found**: Risorsa non trovata
...
    
    ## Tipi di Utente
    
...
    
...
    
...

...
servers:
...
paths:
...
  /api/auth/google:
    post:
      tags: [Authentication]
      operationId: googleLogin
      summary: Login con Google OAuth
      description: Autentica un utente utilizzando il token ID di Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idToken
              properties:
                idToken:
                  type: string
                  description: "Token ID di Google"
                  example: "eyJhbGciOiJSUzI1NiIs..."
      responses:
        '200':
          description: Login Google effettuato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    email: "mario.rossi@gmail.com"
                  isNewUser: false
                message: "Google login successful"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/link-google:
    post:
      tags: [Authentication]
      operationId: linkGoogleAccount
      summary: Collega account Google esistente
      description: |
        Collega un account Google a un account esistente nel sistema.
        Permette login futuro tramite Google OAuth.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idToken
              properties:
                idToken:
                  type: string
                  description: "Token ID di Google"
                  example: "eyJhbGciOiJSUzI1NiIs..."
      responses:
        '200':
          description: Account Google collegato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  googleEmail: "mario.rossi@gmail.com"
                message: "Google account linked successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Account Google già collegato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Google account already linked"
                error:
                  code: "GOOGLE_ALREADY_LINKED"
                  details: "This Google account is already linked to another user"
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/password:
    patch:
      tags: [Authentication]
      operationId: updatePassword
      summary: Aggiorna password utente
      description: |
        Modifica la password dell'utente autenticato.
        Disponibile per tutti i tipi di utente (privato, ente, operatore).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
                  minLength: 8
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
                  description: "Nuova password (minimo 8 caratteri, 1 maiuscola, 1 minuscola, 1 numero e 1 carattere speciale)"
                  example: "NewSecurePass123!"
                currentPassword:
                  type: string
                  description: "Password attuale (richiesta per sicurezza)"
                  example: "OldSecurePass123!"
      responses:
        '200':
          description: Password aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Password updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =============================================================================
  # USER ENDPOINTS (UNIFICATI: privati + enti)
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token ottenuto tramite login"
  responses:
    BadRequest:
      description: "Richiesta non valida - dati mancanti o formato errato"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Invalid request data"
            error:
              code: "BAD_REQUEST"
              details: "The provided data is not valid"
    Unauthorized:
      description: "Non autorizzato - token mancante o non valido"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Authentication required"
            error:
              code: "UNAUTHORIZED"
              details: "Token is missing or invalid"
...
    NotFound:
      description: "Risorsa non trovata"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Resource not found"
            error:
              code: "NOT_FOUND"
              details: "The requested resource does not exist"
...
    InternalError:
      description: "Errore interno del server"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Internal server error"
            error:
              code: "INTERNAL_SERVER_ERROR"
              details: "An unexpected error occurred"
  schemas:
...
    StandardResponse:
      type: object
      properties:
        data:
          type: object
          description: "Contenuto della risposta (opzionale in caso di errore)"
        message:
          type: string
          description: "Messaggio informativo sempre presente"
          example: "Operation completed successfully"
        error:
          type: object
          description: "Dettagli errore (solo in caso di errore)"
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "Detailed error description"
      required:
        - message

    # =============================================================================
    # USER SCHEMA (UNIFICATO)
    # =============================================================================
...
security:
  - BearerAuth: []
  - {}